Set menu as default tab
