+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "restaurant-page",
   "version": "1.0.0",
-  "description": "This project dynamically renders the home page for a restaurant's website",
+  "description": "This project renders the home page for a restaurant's website",
   "private": true,
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
