+++ src/js/dom.js
@@ -1,4 +1,24 @@
-import { getTabContent } from './tab';
+import about from './pages/about';
+import contact from './pages/contact';
+import menu from './pages/menu';
+
+const getTabContent = (tabName) => {
+  let tabContent;
+  switch (tabName) {
+    case 'menu':
+      tabContent = menu;
+      break;
+    case 'about':
+      tabContent = about;
+      break;
+    case 'contact':
+      tabContent = contact;
+      break;
+    default:
+      tabContent = 'ERROR!! Tab name not valid';
+  }
+  return tabContent;
+};
 
 const writeTabContents = (htmlContent) => {
   const activeTabContents = document.getElementById('tab-content');
@@ -6,19 +26,24 @@ const writeTabContents = (htmlContent) => {
 };
 
 const clearTabs = () => {
-  writeTabContents('');
   const tabs = [...document.getElementsByClassName('tab-heading')];
   tabs.forEach(tab => {
-    tab.style.border = 'none';
+    if (tab.classList.contains('active-tab')) {
+      tab.classList.remove('active-tab');
+    }
   });
+  writeTabContents('');
 };
 
+const activateTab = (tab) => {
+  tab.classList.toggle('active-tab');
+  writeTabContents(getTabContent(tab.id));
+};
 
 const switchTab = (event) => {
   clearTabs();
   const tabButton = event.target.closest('div');
-  tabButton.style.borderBottom = '2px solid red';
-  writeTabContents(getTabContent(tabButton.id));
+  activateTab(tabButton);
 };
 
-export default switchTab;
+export { switchTab, activateTab };
